Refactor: 브렌치 초기값으로 복구
